unselected courses added as desired
