+++ src/variables.scss
@@ -33,6 +33,8 @@ $modal-strong-foreground-color-light: #444444;
 $seat-info-dark-label: rgba(255, 255, 255, 0.50);
 $seat-info-light-label: #848173;
 
+$course-page-link-color: #8bd6fb;
+
 @mixin card {
   box-shadow: 0 1px 3px 0 rgba(0, 0, 0, 0.2), 0 1px 1px 0 rgba(0, 0, 0, 0.14),
     0 2px 1px -1px rgba(0, 0, 0, 0.12);
